requirements.txt
fastapi~=0.115.4
uvicorn~=0.32.0
pydantic~=2.9.2
sqlalchemy~=2.0.36
asyncpg~=0.30.0
alembic~=1.13.3
pillow~=11.0.0
aio_pika~=9.4.3
python-dotenv~=1.0.1
jwt~=1.3.1
psycopg2-binary~=2.9.10
python-multipart~=0.0.16
passlib[bcrypt]~=1.7.4

